WIP, more code refactoring - extract arrow flight service calls out of Session
